+++ container/container_test.go
@@ -35,7 +35,7 @@ func mockNew(m *mocks.CommonAPIClient) {
 		}
 		networkInspectArguments = []interface{}{
 			mock.Anything,
-			"core",
+			"engine",
 			types.NetworkInspectOptions{},
 		}
 		networkInspectResponse = types.NetworkResource{
@@ -117,7 +117,7 @@ func TestNew(t *testing.T) {
 	}
 
 	ln := <-dt.LastNetworkCreate()
-	require.Equal(t, cfg.Core.Name, ln.Name)
+	require.Equal(t, cfg.Name, ln.Name)
 	require.Equal(t, types.NetworkCreate{
 		CheckDuplicate: true,
 		Driver:         "overlay",