+++ service/service.go
@@ -53,6 +53,9 @@ type Service struct {
 	// a Git host.
 	Repository string `hash:"name:7"`
 
+	// Source is the hash id of service's source code on IPFS.
+	Source string `hash:"name:9"`
+
 	// DeployedAt holds the creation time of service.
 	DeployedAt time.Time `hash:"-"`
 }
@@ -91,8 +94,8 @@ type Log struct {
 	Error      io.ReadCloser
 }
 
-// getDependency returns dependency dependencyKey or a not found error.
-func (s *Service) getDependency(dependencyKey string) (*Dependency, error) {
+// GetDependency returns dependency dependencyKey or a not found error.
+func (s *Service) GetDependency(dependencyKey string) (*Dependency, error) {
 	for _, dep := range s.Dependencies {
 		if dep.Key == dependencyKey {
 			return dep, nil